First complete version.
Very barebones. Card database very incomplete and only contains around 200 cards. Only like 11186 more to go...
Asks you if you want it to output the entire current database. After that it reads all of the files that are included in this version to test the program.
Card IDs that aren't in the database will be stored and output at the end of the console.

Decks used:
https://ygoprodeck.com/trickstar-23/
https://ygoprodeck.com/anime-format-trickstar-shaddoll/
https://ygoprodeck.com/trickstar-24/
https://ygoprodeck.com/trickstar-post-tcg-banlist/
https://ygoprodeck.com/farstar/
https://ygoprodeck.com/trickstar-post-tcg-banlist-may-2022/
https://ygoprodeck.com/idol-queens-trickstar-deck/
https://ygoprodeck.com/dont-play-this-damn-it/
https://ygoprodeck.com/slow-roast-idols-v2/
https://ygoprodeck.com/burning-trickstar/
https://ygoprodeck.com/trickstar-master-duel/
https://ygoprodeck.com/enforcer-trickstars/
https://ygoprodeck.com/trickstar-adventurer-p-u-n-k-eldlich/